Got as far as I could before realizing the mini-server won't work
At least it runs, but it won't matter as long as CORS is a browser level problem. That means the browser itself needs to be bypassed somehow. The mini-server needs to be a background process that is automatically deployed by the dApp. That means the dApp has to make the request for such an installation and run.